Changes to Layout and minor changes

+++ ProjectGecko/Models/Comment.cs
@@ -7,7 +7,7 @@ namespace ProjectGecko.Models
 {
     public class Comment
     {
-        public int CommentPoster { get; set; }
+        public long CommentPoster { get; set; }
         public string Text { get; set; }
     }
 }
